+++ src/client/react.js
@@ -105,7 +105,7 @@ export function Client({
       if (board) {
         _board = React.createElement(board, {
           ...state,
-          isMultiplayer: multiplayer === true,
+          isMultiplayer: multiplayer !== undefined,
           moves: this.client.moves,
           events: this.client.events,
           gameID: this.props.gameID,
@@ -121,7 +121,7 @@ export function Client({
         _debug = React.createElement(Debug, {
           gamestate: state,
           store: this.client.store,
-          isMultiplayer: multiplayer === true,
+          isMultiplayer: multiplayer !== undefined,
           moves: this.client.moves,
           events: this.client.events,
           gameID: this.props.gameID,